pyproject.toml
[project]
name = "distributed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logfire>=4.16.0",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
